java/src/main/java/com/yourtion/leetcode/daily/m10/d25/Solution.java
package com.yourtion.leetcode.daily.m10.d25;

/**
 * 845. 数组中的最长山脉
 *
 * @author Yourtion
 * @link https://leetcode-cn.com/problems/longest-mountain-in-array/
 */
public class Solution {
    public int longestMountain(int[] A) {
        int n = A.length;
        int ans = 0;
        int left = 0;
        while (left + 2 < n) {
            int right = left + 1;
            if (A[left] < A[left + 1]) {
                while (right + 1 < n && A[right] < A[right + 1]) {
                    ++right;
                }
                if (right < n - 1 && A[right] > A[right + 1]) {
                    while (right + 1 < n && A[right] > A[right + 1]) {
                        ++right;
                    }
                    ans = Math.max(ans, right - left + 1);
                } else {
                    ++right;
                }
            }
            left = right;
        }
        return ans;
    }
}
